requirements_lock.txt
# Generated package requirements for local venv
# Do not edit manually - this file is generated by Bazel

pytest==8.3.5
pydantic==2.11.3
typer==0.15.2
httpx==0.28.1
pyyaml==6.0.2
ruff==0.11.6
typing-extensions==4.11.0